Pipeline refactor: Refactor services module into package
