+++ data_processors/pipeline/lambdas/germline.py
@@ -3,29 +3,24 @@ try:
 except ImportError:
     pass
 
-import django
 import os
+import django
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'data_portal.settings.base')
 django.setup()
 
 # ---
 
-# Standards
 import copy
 import logging
-from datetime import datetime, timezone
 
-# Data portal imports
 from data_portal.models import Workflow
-from data_processors.pipeline import services
+from data_processors.pipeline.services import sequence_srv, batch_srv, workflow_srv
 from data_processors.pipeline.constant import WorkflowType, WorkflowHelper
 from data_processors.pipeline.lambdas import wes_handler
 
-# Utils imports
 from utils import libjson, libssm, libdt
 
-# Set loggers
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
@@ -119,10 +114,10 @@ def handler(event, context) -> dict:
     workflow_id = libssm.get_ssm_param(wfl_helper.get_ssm_key_id())
     workflow_version = libssm.get_ssm_param(wfl_helper.get_ssm_key_version())
 
-    sqr = services.get_sequence_run_by_run_id(seq_run_id) if seq_run_id else None
-    batch_run = services.get_batch_run(batch_run_id=batch_run_id) if batch_run_id else None
+    sqr = sequence_srv.get_sequence_run_by_run_id(seq_run_id) if seq_run_id else None
+    batch_run = batch_srv.get_batch_run(batch_run_id=batch_run_id) if batch_run_id else None
 
-    matched_runs = services.search_matching_runs(
+    matched_runs = workflow_srv.search_matching_runs(
         type_name=WorkflowType.GERMLINE.name,
         wfl_id=workflow_id,
         version=workflow_version,
@@ -168,7 +163,7 @@ def handler(event, context) -> dict:
         'workflow_input': workflow_input,
     }, context)
 
-    workflow: Workflow = services.create_or_update_workflow(
+    workflow: Workflow = workflow_srv.create_or_update_workflow(
         {
             'wfr_name': workflow_run_name,
             'wfl_id': workflow_id,
